Reduce big structural duplication
Three behavior-preserving structural refactors. No logic changes; the large
deletion count is duplicated/boilerplate code removed.

- GameViewModel: add observePerCharacter {} (replaces 8 hand-rolled
  characterId.flatMapLatest { ... else flow {} } flows) and observeCharacterInt
  (collapses the 4 identical dimension flows), plus a createWindowData factory
  that replaces the when(windowType) -> Stream/Dialog/null block copy-pasted in
  three places.

- WindowView / DesktopWindowView: move the 9 byte-identical helpers
  (WindowBackgroundImage, backgroundAlignment, contentScale, gradient/opacity
  math, toPercentFraction, isShowing, isPreviousPrompt) into a shared
  commonMain WindowViewHelpers.kt; ~180 lines of duplication removed from each
  file. The platform files keep only their toolkit-specific UI.

- Settings views: extract SettingsListScaffold (one mobile/Material3, one
  desktop/Jewel) for the repeated character-selector + title preamble, applied
  across all 12 settings list screens. Per-view list/buttons are passed as a
  ColumnScope content slot, so every per-view difference (bottom bars,
  modifiers, allowGlobal) is preserved verbatim.

Verified: compose (JVM + Android), desktopApp, and androidApp compile; ktlint
clean. iOS not built locally (needs macOS), but no iOS-specific code changed.
