+++ compose/src/jvmMain/kotlin/warlockfe/warlock3/compose/desktop/ui/settings/DesktopNamesView.kt
@@ -57,16 +57,13 @@ fun DesktopNamesView(
     var editingName by remember { mutableStateOf<NameEntity?>(null) }
     val scope = rememberCoroutineScope()
 
-    Column(modifier.fillMaxSize()) {
-        DesktopSettingsCharacterSelector(
-            selectedCharacter = selectedCharacter,
-            characters = allCharacters,
-            onSelect = { selectedCharacter = it },
-            allowGlobal = true,
-        )
-        Spacer(Modifier.height(16.dp))
-        Text("Names")
-        Spacer(Modifier.height(8.dp))
+    SettingsListScaffold(
+        title = "Names",
+        selectedCharacter = selectedCharacter,
+        characters = allCharacters,
+        onSelectCharacter = { selectedCharacter = it },
+        modifier = modifier.fillMaxSize(),
+    ) {
         WarlockScrollableColumn(modifier = Modifier.fillMaxWidth().weight(1f)) {
             names.forEach { name ->
                 WarlockListItem(
