+++ compose/src/jvmMain/kotlin/warlockfe/warlock3/compose/desktop/ui/settings/SettingsListScaffold.kt
@@ -1,0 +1,40 @@
+package warlockfe.warlock3.compose.desktop.ui.settings
+
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.height
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import org.jetbrains.jewel.ui.component.Text
+import warlockfe.warlock3.core.client.GameCharacter
+
+/**
+ * The shared preamble for the desktop per-character settings list screens: a character selector
+ * followed by a title. Each screen supplies its own list and action buttons as [content] (a
+ * [ColumnScope] so the list can take the remaining height with `Modifier.weight(1f)`).
+ */
+@Composable
+fun SettingsListScaffold(
+    title: String,
+    selectedCharacter: GameCharacter?,
+    characters: List<GameCharacter>,
+    onSelectCharacter: (GameCharacter?) -> Unit,
+    modifier: Modifier = Modifier,
+    allowGlobal: Boolean = true,
+    content: @Composable ColumnScope.() -> Unit,
+) {
+    Column(modifier) {
+        DesktopSettingsCharacterSelector(
+            selectedCharacter = selectedCharacter,
+            characters = characters,
+            onSelect = onSelectCharacter,
+            allowGlobal = allowGlobal,
+        )
+        Spacer(Modifier.height(16.dp))
+        Text(title)
+        Spacer(Modifier.height(8.dp))
+        content()
+    }
+}
