+++ compose/src/mobileMain/kotlin/warlockfe/warlock3/compose/ui/window/WindowView.kt
@@ -499,164 +499,6 @@ private fun WindowViewContent(
 }
 
 @Composable
-private fun WindowBackgroundImage(
-    backgroundImage: ClientBackgroundImage,
-    width: Dp,
-    height: Dp,
-    modifier: Modifier = Modifier,
-) {
-    val painter =
-        rememberAsyncImagePainter(
-            ImageRequest
-                .Builder(LocalPlatformContext.current)
-                .data(backgroundImage.image)
-                .size(Size.ORIGINAL)
-                .build(),
-        )
-    val painterState by painter.state.collectAsState()
-    val state = painterState as? AsyncImagePainter.State.Success ?: return
-    val imageHeight =
-        state.result.image.height
-            .takeIf { it > 0 } ?: return
-    val imageWidth =
-        state.result.image.width
-            .takeIf { it > 0 } ?: return
-    val aspectRatio = imageWidth.toFloat() / imageHeight.toFloat()
-    val imageWidthDp = with(LocalDensity.current) { imageWidth.toDp() }
-    val imageHeightDp = with(LocalDensity.current) { imageHeight.toDp() }
-    val scaledWidth = height * aspectRatio
-    val scaledHeight = width / aspectRatio
-    val imageModifier =
-        when (backgroundImage.mode) {
-            BackgroundImageMode.FILL -> {
-                modifier.fillMaxSize()
-            }
-
-            BackgroundImageMode.HEIGHT_FILL,
-            BackgroundImageMode.GRADIENT,
-            -> {
-                modifier
-                    .fillMaxHeight()
-                    .width(scaledWidth)
-            }
-
-            BackgroundImageMode.WIDTH_FILL -> {
-                modifier
-                    .fillMaxWidth()
-                    .height(scaledHeight)
-            }
-
-            BackgroundImageMode.FULL -> {
-                modifier
-                    .requiredWidth(imageWidthDp)
-                    .requiredHeight(imageHeightDp)
-            }
-        }.then(backgroundImage.opacityModifier())
-            .then(backgroundImage.gradientModifier())
-
-    Image(
-        modifier = imageModifier,
-        painter = painter,
-        contentDescription = null,
-        contentScale = backgroundImage.mode.contentScale(),
-    )
-}
-
-private fun ClientBackgroundImage.backgroundAlignment(): Alignment =
-    when (verticalAlignment) {
-        BackgroundImageVerticalAlignment.TOP -> {
-            when (horizontalAlignment) {
-                BackgroundImageHorizontalAlignment.LEFT -> Alignment.TopStart
-                BackgroundImageHorizontalAlignment.CENTER -> Alignment.TopCenter
-                BackgroundImageHorizontalAlignment.RIGHT -> Alignment.TopEnd
-            }
-        }
-
-        BackgroundImageVerticalAlignment.MIDDLE -> {
-            when (horizontalAlignment) {
-                BackgroundImageHorizontalAlignment.LEFT -> Alignment.CenterStart
-                BackgroundImageHorizontalAlignment.CENTER -> Alignment.Center
-                BackgroundImageHorizontalAlignment.RIGHT -> Alignment.CenterEnd
-            }
-        }
-
-        BackgroundImageVerticalAlignment.BOTTOM -> {
-            when (horizontalAlignment) {
-                BackgroundImageHorizontalAlignment.LEFT -> Alignment.BottomStart
-                BackgroundImageHorizontalAlignment.CENTER -> Alignment.BottomCenter
-                BackgroundImageHorizontalAlignment.RIGHT -> Alignment.BottomEnd
-            }
-        }
-    }
-
-private fun BackgroundImageMode.contentScale(): ContentScale =
-    when (this) {
-        BackgroundImageMode.FILL -> ContentScale.FillBounds
-
-        BackgroundImageMode.WIDTH_FILL -> ContentScale.FillWidth
-
-        BackgroundImageMode.FULL -> ContentScale.None
-
-        BackgroundImageMode.HEIGHT_FILL,
-        BackgroundImageMode.GRADIENT,
-        -> ContentScale.FillHeight
-    }
-
-private fun ClientBackgroundImage.gradientModifier(): Modifier =
-    when (mode) {
-        BackgroundImageMode.GRADIENT -> {
-            Modifier
-                .graphicsLayer {
-                    compositingStrategy = CompositingStrategy.Offscreen
-                }.drawWithContent {
-                    drawContent()
-                    drawRect(
-                        brush = Brush.horizontalGradient(*gradientColorStops()),
-                        blendMode = BlendMode.DstIn,
-                    )
-                }
-        }
-
-        else -> {
-            Modifier
-        }
-    }
-
-private fun ClientBackgroundImage.gradientColorStops(): Array<Pair<Float, Color>> {
-    val start = gradientStart.toPercentFraction()
-    val end = gradientEnd.toPercentFraction()
-    val transparent = Color.Black.copy(alpha = 0f)
-    val opaque = Color.Black.copy(alpha = opacity.toPercentFraction())
-
-    return if (start <= end) {
-        arrayOf(
-            0f to transparent,
-            start to transparent,
-            end to opaque,
-            1f to opaque,
-        )
-    } else {
-        arrayOf(
-            0f to opaque,
-            end to opaque,
-            start to transparent,
-            1f to transparent,
-        )
-    }
-}
-
-private fun ClientBackgroundImage.opacityModifier(): Modifier =
-    if (mode == BackgroundImageMode.GRADIENT || opacity == 100) {
-        Modifier
-    } else {
-        Modifier.graphicsLayer {
-            alpha = opacity.toPercentFraction()
-        }
-    }
-
-private fun Int.toPercentFraction(): Float = coerceIn(0, 100) / 100f
-
-@Composable
 private fun ActionContextMenu(
     offset: Offset?,
     menuData: WarlockMenuData,
@@ -754,25 +596,4 @@ private fun ActionContextMenu(
             }
         }
     }
-}
-
-private fun StreamTextLine.isShowing(openWindows: List<String>): Boolean =
-    this.showWhenClosed == null ||
-        openWindows.none {
-            it == this.showWhenClosed
-        }
-
-private fun List<StreamLine>.isPreviousPrompt(
-    index: Int,
-    openWindows: List<String>,
-): Boolean {
-    var currentIndex = index - 1
-    while (currentIndex >= 0) {
-        val line = this[currentIndex] as? StreamTextLine ?: return false
-        if (line.isShowing(openWindows)) {
-            return line.isPrompt
-        }
-        currentIndex--
-    }
-    return false
 }
